website versions/version2/site_/requirements.txt
django
gunicorn
django-heroku
dj-database-url
dj-static
django-taggit
Unipath
python-decouple
Pillow
Markdown
bleach
psycopg2
whitenoise
django-hitcount
